Moved some of roulette functions to different file
